+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 
 ## [Unreleased]
 ### Added
+- Implemented `is_mud_strider` trait. Entities with this trait consume 0 stamina when moving on mud terrain and gain a defense bonus while on mud.
 - Fixed missing mutation logic and added missing visualizer tests for `is_grass_strider`.
 - Fixed missing mutation logic for `is_space_strider`.
 - Implemented `is_storm_strider` trait. Entities with this trait consume 0 stamina when moving during a 'storm' event and gain a defense bonus.